watcher/requirements.txt
requests
pandas
pyyaml
pytest